Introduce new command parsing system
BasilC commands will now be parsed through the new parser found in
cmd.c. Currently only BasilC-end() is migrated, but eventually all
parsing code will be migrated to the new system. To be parsed in the
new system, commands must be declared in the libbasilc source and
include folder (see end() for examples). New commands must also be
added to libbasilc.c in the libbasilc_register() method in order
for them to be recognized by the parser.